make config polling independent from tauri
