--- src-tauri/.sqlx/query-2d6a90bf82aa92118b36d90bf16ae9515952d077106d303f398167caf68a8a70.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "SELECT count(*) FROM tunnel_stats",
-  "describe": {
-    "columns": [
-      {
-        "name": "count(*)",
-        "ordinal": 0,
-        "type_info": "Integer"
-      }
-    ],
-    "parameters": {
-      "Right": 0
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "2d6a90bf82aa92118b36d90bf16ae9515952d077106d303f398167caf68a8a70"
-}
