+++ src-tauri/enterprise/config-sync/Cargo.toml
@@ -18,7 +18,6 @@ semver = { version = "1.0", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sqlx = { version = "0.8", features = ["chrono", "runtime-tokio", "sqlite"] }
-tauri = { version = "2", features = ["tray-icon"] }
 tokio = { version = "1", features = ["time"] }
 
 [target.'cfg(not(target_os = "macos"))'.dependencies]
